Move getParameter() API to function type
