+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/api/types/ResourceMethodType.java
@@ -17,8 +17,6 @@
  */
 package io.ballerina.runtime.api.types;
 
-import io.ballerina.runtime.api.Parameter;
-
 /**
  * {@code ResourceFunctionType} represents a resource function in Ballerina.
  *
@@ -31,5 +29,4 @@ public interface ResourceMethodType extends MethodType {
     String[] getResourcePath();
     @Deprecated
     Boolean[] getParamDefaultability();
-    Parameter[] getParameters();
 }
