+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/internal/types/BMethodType.java
@@ -20,6 +20,7 @@ package io.ballerina.runtime.internal.types;
 import io.ballerina.runtime.api.types.FunctionType;
 import io.ballerina.runtime.api.types.MethodType;
 import io.ballerina.runtime.api.types.ObjectType;
+import io.ballerina.runtime.api.types.Parameter;
 import io.ballerina.runtime.api.types.Type;
 import io.ballerina.runtime.api.utils.IdentifierUtils;
 
@@ -41,6 +42,11 @@ public class BMethodType extends BFunctionType implements MethodType {
         this.type = type;
         this.parentObjectType = parent;
         this.flags = flags;
+    }
+
+    public BMethodType(String funcName, BObjectType parent, BFunctionType type, long flags, Parameter[] parameters) {
+        this(funcName, parent, type, flags);
+        this.parameters = parameters;
     }
 
     @Override
@@ -78,6 +84,6 @@ public class BMethodType extends BFunctionType implements MethodType {
     }
 
     public <T extends MethodType> MethodType duplicate() {
-        return new BMethodType(funcName, parentObjectType, type, flags);
+        return new BMethodType(funcName, parentObjectType, type, flags, parameters);
     }
 }
